refactor(inventory): Phase 4 split EfInventoryStore from InventoryContext
- New EfInventoryStore in Infrastructure/Data/EntityFramework/ implements
  IInventoryStore via ctx.<DbSet>; DI flipped to resolve EfInventoryStore.
- InventoryContext reduced to DbContext + DbSets + OnModelCreating +
  RecordStockMovement (visibility relaxed to internal so EfInventoryStore
  can record movements through the context-owned metric helper).
